rename: 엔티티 패키지 구조 개편 및 import 경로 수정
