+++ src/main/java/com/sprint/mission/discodeit/DiscodeitApplication.java
@@ -1,22 +1,12 @@
 package com.sprint.mission.discodeit;
 
-import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.Message;
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.entity.main.Channel;
+import com.sprint.mission.discodeit.entity.main.Message;
+import com.sprint.mission.discodeit.entity.main.User;
 import com.sprint.mission.discodeit.entity.enums.ChannelType;
-import com.sprint.mission.discodeit.repository.ChannelRepository;
-import com.sprint.mission.discodeit.repository.MessageRepository;
-import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.repository.file.FileChannelRepository;
-import com.sprint.mission.discodeit.repository.file.FileMessageRepository;
-import com.sprint.mission.discodeit.repository.file.FileUserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.service.UserService;
-import com.sprint.mission.discodeit.service.basic.BasicChannelService;
-import com.sprint.mission.discodeit.service.basic.BasicMessageService;
-import com.sprint.mission.discodeit.service.basic.BasicUserService;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -37,8 +27,8 @@ public class DiscodeitApplication {
         MessageService messageService = context.getBean(MessageService.class);
 
         userCRUDTest(userService);
-        channelCRUDTest(channelService);
-        messageCRUDTest(userService, channelService, messageService);
+//        channelCRUDTest(channelService);
+//        messageCRUDTest(userService, channelService, messageService);
     }
 
     private static void userCRUDTest(UserService userService) {
