+++ src/main/java/com/sprint/mission/discodeit/entity/main/Message.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.main;
 
 import lombok.Getter;
