+++ src/main/java/com/sprint/mission/discodeit/repository/ChannelRepository.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.repository;
 
-import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.main.Channel;
 
 import java.util.List;
 import java.util.Optional;
