+++ src/main/java/com/sprint/mission/discodeit/repository/MessageRepository.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.repository;
 
-import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.entity.main.Message;
 
 import java.util.List;
 import java.util.Optional;
