+++ src/main/java/com/sprint/mission/discodeit/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.repository;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.entity.main.User;
 
 import java.util.List;
 import java.util.Optional;
