+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFUserRepository.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.repository.jcf;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.entity.main.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 
 import java.util.*;
