+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.main.Channel;
 import com.sprint.mission.discodeit.entity.enums.ChannelType;
 
 import java.util.List;
