+++ src/main/java/com/sprint/mission/discodeit/service/MessageService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.entity.main.Message;
 
 import java.util.List;
 import java.util.UUID;
