+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFUserService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service.jcf;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.entity.main.User;
 import com.sprint.mission.discodeit.service.UserService;
 
 import java.util.*;
